% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exData}
\alias{exData}
\title{Example Continuous Response ANOVA Dataset.}
\format{
A matrix with 80 rows and 14 columns:
\describe{
\item{Column 1}{Continuous Response Variable}
\item{Columns 2-14}{Design Matrix}
}
}
\source{
(Not sure, need to update)
}
\usage{
exData
}
\description{
A dataset containing an example test set result to run a Predictive Probability model.
(Add more description)
}
\keyword{datasets}
