% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.ContRespPP}
\alias{print.ContRespPP}
\title{Print method for ContRespPP class}
\usage{
\method{print}{ContRespPP}(x, ...)
}
\arguments{
\item{x}{A list of class ContRespPP}

\item{...}{Other arguments passed to \code{print}}
}
\value{
Prints the predictive or posterior probability result from ContRespPP class
}
\description{
Print method for ContRespPP class
}
