% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.ContRespPP}
\alias{summary.ContRespPP}
\title{Summary method for ContRespPP class}
\usage{
\method{summary}{ContRespPP}(object, ...)
}
\arguments{
\item{object}{A list of class ContRespPP}

\item{...}{Other arguments passed to \code{summary}}
}
\value{
Prints a basic summary of posterior distribution from ContRespPP class
}
\description{
Summary method for ContRespPP class
}
